Fix: Solve problems in crdma_query_pkey
After crdma driver loaded, we use the dmesg command
to see that crdma_query_pkey message is failed. In
the crdma_query_pkey function, we should use index
instead of port_num in the if condition. We also
use CRDMA_IB_MAX_PKEY_TABLE_SIZE instead of
CRDMA_IB_MAX_PKEY for better representation. It is
used to initialize max_pkeys of device and
pkey_tbl_len of port.

Signed-off-by: Jiaqi Han <jiaqi.han@corigine.com>
Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Signed-off-by: Huibin Li <huibin.li@corigine.com>