GUID: Use addrconf_addr_eui48 to replace crdma_mac_to_guid
In RDMA, we need to create GUID using MAC of interface.
By now we implement this function with crdma_mac_to_guid,
but in kernel a general API addrconf_addr_eui48 can do this,
so we replace it to keep consistency with community.
Remove crdma_mac_to_guid which is unused.

Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Huibin Li <huibin.li@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Signed-off-by: Xiaofeng Tang <xiaofeng.tang@corigine.com>
Signed-off-by: Fangzhou Feng <fangzhou.feng@corigine.com>