QP: Add the IB_QP_CUR_STATE when modifying QP
Add IB_QP_CUR_STATE to our code to solve any inconsistencies
between the current state of the qp and the user-specified
state.
When an error occurs in QP, the hardware will automatically
convert the status of QP to ERR, but the status of QP is not
updated inside the driver. We can use IB_QP_CUR_STATE to hint
the stack what the current QP status is.
However, our microcode does not support the automatic conversion
of QP state so far, and the conversion of QP state is carried
out in the driver. So IB_QP_CUR_STATE is added to prepare for
the inconsistency between the current state of qp and the state
specified by the user in the future. Now we only support partially,
this part needs to be adapted latter.

Signed-off-by: Jiaqi Han <jiaqi.han@corigine.com>
Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Signed-off-by: Huibin Li <huibin.li@corigine.com>