Compat: cmd_task compatible with different kernel versions
Convert the write interface to use the uverbs_api in linux kernel
versions 5.0 and above. Set the write command as an object and
link to uverbs_api, so the command path is redone to use uapi.
The command mask function(verify_command_mask) is replaced by the
registration time check in the uapi.

In the kernel above 5.11, The _ib_alloc_device function in the
core code presets a list of uversb_cmd_masks.so we remove element
from uversb_cmd_mask in the crdma_verbs.c file.

Signed-off-by: Chenchen Liu <chenchen.liu@corigine.com>
Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Signed-off-by: Huibin Li <huibin.li@corigine.com>