nfp: doorbell: add support for hardware doorbell
Start with Kestrel, nfp begin to support hardware doorbell,
it is used to raise signal from host driver to nfp with
information. RoCE and VDPA features need this feature,
so we extract it as new feature. In this commit we add
initialization of doorbell, include reserving PCIE bar,
adaption for bar matching workflow which helps driver
to get cpp area from doorbell bar space.

Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>