nfp: add capability and control flags for metadata padding
Add capability and control flags to support the addition of
metadata padding for the purposes of DMA alignment. Since
metadata padding is to be added in the driver TX datapath,
it needs to be stripped by firmware. Support for this
functionality must be relayed from firmware to driver for
compatibility.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>