DSCP/ECN: Support carrying DSCP/ECN to roce packet
The dscp field addition process already exists in our software,
but firmware fetch the wrong value, as the 'traffic_class' value
passed from driver has been (<<2) yet.

Now also in order to support ecn in roce packet, in driver we
fill 'traffic_class' value passed to firmware with full 8-bits
value composed of 6-bits dscp field and 2-bits ecn field.

The 'traffic_class' value is part of the AH which contains all
of the necessary data to reach a remote destination.

For RC, the AH is associated with a queue pair (QP), and is passed
to firmware when qp status is modified to RTR with IBV_QP_AV attr.
Driver can fill ecn value when it happend.

For UD, the AH is associated with a work request (WR), RDMA
applications call uverb 'ibv_create_ah' to create AH, and then use
this AH for the WR. So, here driver can return the 'traffic_class'
value filled with ecn to RDMA applications for the transport.

Signed-off-by: Li Xu <li.xu@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Signed-off-by: Xiaofeng Tang <xiaofeng.tang@corigine.com>
Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Xinxia Qu <xinxia.qu@corigine.com>