nfp: nfdk tx update stats counter to exclude metadata bytes
Subtract metadata size from tx statistics.

Checks whether the metadata descriptor indicates that metadata was
prepended. In such cases, read the number of metadata bytes from the
front of the skb and subtract it from the stats.

This will underestimate the statistics in certain scenarios, such as when
vlan tags are offloaded. This aligns with the firmware's stats tracking.

Signed-off-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>