nfp: flower: bind 2 port to pf0 using multiple PF BSP
Apply disa firmware and driver on BSP with multiple-PF enabled. Load PF0
with 2 ports bound as single PF version. Skip loading of PF1.

Trick is to read app_id for PF0 instead of PF1 because resource for PF1 is
actually not allocated. So adding a helper to read app_id for a specific PF
(PF0).

Signed-off-by: Tianyu.Yuan <tianyu.yuan@corigine.com>