nfp: debugfs: add cpp userspace toggle
Add a debugfs file to enable external cpp access to the user. It
implements a read and write handle - read will return 0 or 1 depending
on whether or not it is exposed. And the write callback will enable
or disable it depending on whether or not a 1 or 0 was written to it.

Also include the required compat code for kstrtobool_from_user().

Signed-off-by: Louis Peens <louis.peens@corigine.com>