Bugfix: Support SQ signaling all of QP in kernel driver
Support SQ signaling all of QP in kernel driver. The reason
we use sq_sig_all not sq_sig_type are:
1. Reference to other vendor providers
2. Keep correspondence with userspace implement
If QP is created with sq_sig_all set, we will set signal flag
for each SWQE of this QP. For firmware side, only signaled
SWQE needs to be processed and sq_sig_all of QP is not perceived.

Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Signed-off-by: Xiaofeng Tang <xiaofeng.tang@corigine.com>