packaging: rpm: remove use of find-provides
With the release of dkms v3.0.12, the /usr/lib/dkms/find-provides script
is no longer available and this resulted in the building of the driver
package to fail. The commit: 046741f ("Remove find-provides script")
in the dkms repository removed the find-provides script because it was
only used for `make install-redhat` but it has since been removed.

Since this script is no longer available and it appears to have no
affect on the packages that gets built, it is removed from the rpmbuild
arguments.

Signed-off-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>