nfp: fix issues of vfs on pf1
There are link and traffic issues for vfs on pf1. When generating
port id for vf repr, we need pcie port id, it is set to pf id
previously, but it need to be set to 0. Also for vnic id of vf repr,
we need to pass absolute value, so for vfs on pf1, this value should
start with the vf offset defined in pcie sriov capability.

Signed-off-by: Baowen Zheng <baowen.zheng@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>