nfp: avoid padding too long metadata in LSO case
Since only 8 bits are defined for header length in LSO descriptor,
and the header length includes prepended metadata if existing. And
the padding metadata in some environment whose cache line is 128B
aligned makes it possible that the final header length exceeds 255
bytes in some encapsulated and IPv6 cases. So add some code to drop
the padding metadata to avoid it happens.

Fixes: a1c36074c43c ("nfp: add nfdk tx metadata padding for alignment")
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>