nfp: flower: offload tun flow only when tun IP present
Only offload flower tunnel flows if the tunnel source IP
is present on an NFP netdev or OVS internal port, preventing
the offloading of flows with incorrect source addresses.
Previously, in the absence of the tunnel IP, the kernel
assigned the default route as source address, rendering the
flow invalid. An offload flag in the neighbour table is used
to keep track of flow offloading.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>