Bugfix: MPT is not cleared when MR is destroyed
In RDMA driver, when MR is destroyed, the index of MPT is
transported to firmware, but firmware uses it as key of MR,
this mismatch leads to an unsuccessful delete. We fix this
with transporting key of MR to delete MPT. The benefit of
key is that firmware can do validation check when MPT
delete is issued.

Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Xiaofeng Tang <xiaofeng.tang@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>