nfp: dcb: add the callback function ".ndo_select_queue" for nfp
When the dcb and pfc function are enabled,this function
can select specific queues for different priority packet
in the driver.

The first step is to retrieve the priority of the packet.
The second step is to select the traffic-class based on priority.
The third step is to select the queue index based on the traffic-class.

Signed-off-by: Weiwen Hu<weiwen.hu@corigine.com>