nfp: flower: offload/del tun flow when tun IP added/removed
Offload, or unoffload, tunnel flows when a tunnel IP address is
added, or removed, from an NFP netdev or OVS internal port.

Flower tunnel flows are only offloaded if the tunnel source IP
is present on an NFP netdev or OVS internal port, preventing the
offloading of flows with incorrect source addresses. If the
tunnel IP address is added after the neighbour entry has been
added, the neighbour table is walked, and matching pending flow
rules are then offloaded.

Conversely, if the tunnel IP address is removed, any matching
already offloaded flows are invalidated and the offload flag
cleared in the neighbour table.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>