nfp: grant the right of reclaiming resources to PF0 only
In multi-PF setup, one PF still may bust the resource lock
that is held by another. Since it's an error handling to
bust lock and we only need a way here, no need to let every
PF have the error handler. Now leave the right to the first
PF only.

Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>