sync: nfp: ethtool: support TX/RX pause frame on/off
Update 'nfp_port_set_pauseparam' to match the version which got
upstreamed. In this case the patch has been present for longer, so a
full revert and re-add is impractical. The only difference is the extra
check in the upstream version to make sure that it is not trying to
enable pause frame autonegotiation. This patch adds in this change.

Signed-off-by: Louis Peens <louis.peens@corigine.com>