workflows: ci-test-dev: remove unnecessary input
It is no longer necessary to specify which branch to use for the driver
source as the nfp-drv-kmods-private `main` branch is now in line with
the release branch, and it contains the multi-pf feature. Tests have
been moved to their own separate branch. The driver source will always
be the nfp-drv-kmods branch that the workflow is run from.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>