packaging: dkms: save module.symvers
The Module.symvers file is needed for external drivers that depend on
the nfp.ko module. Therefore the module-symverse-save.sh script is
added to save/remove the Module.symvers file at
`/lib/modules/${kernel_ver}/extra-symvers/nfp_driver.symvers`. This
location is selected as a location for storing the Module.symvers file
after the driver is built. The external driver's scripts can then point
to this location when needing access the Module.symvers file.

The Makefile is also updated to save the Module.symvers file when
installing the nfp.ko module using the Makefile (make install).

Signed-off-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Signed-off-by: Guibin Xia <guibin.xia@corigine.com>