packaging: get rid of unneeded 'cd $tmpdir' calls
These scripts all follow a similar format where they sometimes store
the current directory into a $tmpdir variable, 'cd' to some place and
then later return back to $tmpdir in preparation for the next function
call. This can be avoided by making the function execute in a sub-shell
instead. This is achieved by using '()' instead of '{}' when defining
the function. This means that any 'cd' commands that happens in the
sub-shell will not impact the path of the main execution flow.

Suggested-by: Walter Heymans <walter.heymans@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>