Feature: Implement Shared Receive Queue
Shared Receive Queue allows multiple Queue Pairs to
share Receive Work Requests and, more importantly, the Data
Segments associated with Receive Work Requests. When an incoming
Receive Message arrives on any QP that is associated with an SRQ,
the HCA uses the next available SRQ WQE to receive the incoming data.
The HCA returns Work Completions through the Completion
Queue that is associated with the QP that received the incoming
Send operation. RC, UC or UD QPs can be associated with an SRQ.

In current version, CRDMA NIC can not get the number of entry in SRQ,
so asynchronous event IBV_EVENT_SRQ_LIMIT_REACHED is not supported.

Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Xiaofeng Tang <xiaofeng.tang@corigine.com>
Signed-off-by: Hui Zhang <hui.zhang@corigine.com>