packaging: avoid modifying the python interpreter of python script
Rpmbuild calls /usr/lib/rpm/redhat/brp-mangle-shebangs to check the
python interpreter if the python script is executable when building
packages on CentOS. The python interpreter of python script may be
modified and the additional dependency '/usr/libexec/platform-python'
can be generated, which is not available on some systems.

This patch disables the execution bit of nfp_troubleshoot_gather on
install stage to ensure the python interpreter is not modified. The
execution bit is set back during post stage.

Signed-off-by: Fei Qin <fei.qin@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>