nfp: fix null-pointer-deref when eth_tbl cannot be read
If something goes wrong when the driver tries to read from the
eth_tbl it just prints out errors, but it does not error stop the
execution path.

Most paths where 'pf->eth_tbl' is used there are safety checks, so
everything still works fine. There is one exception though.

In 'nfp_app_nic_vnic_init_phy_port()' there is a check for an empty
'pf->eth_tbl', but it returns 0, instead of an error code. This causes
'nn->port' to never be allocated, but the execution continues,
since no error is returned. In the end this causes a null pointer
dereference in 'nfp_nic_vnic_init()' when this function is trying to
dereference 'nn->port'.

This patch adds two fixes:
1) Return an error instead of 0 in 'nfp_app_nic_vnic_init_phy_port()'
   when 'pf->eth_tbl == NULL'.
2) Return early in the startup process if getting the eth table
   information failed. There shouldn't be a good reason to continue if
   getting the eth table failed, since many things later on depend on
   the port information being available.

Fix 2) should prevent 1) from happening, but fixing both is probably a
good idea in any case.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>