workflows: package-rpm: new input field driver_upload_location
Add new input field "driver_upload_location" so that the user may
specify where the built RPM should be uploaded to. This will allow users
to run the workflow and target a driver release, but not then overwrite
an existing release RPM, if they just want to upload the final result to
`tmp`.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>