Feature: Implement RoCE Bond feature
This patch realizes the driver part of RoCE bond feature. The feature
mimic Ethernet bonding for RoCE.

The prerequisites for the RoCE bond being active contain:
1. configuration of bond mode should be 1, 2 or 4;
2. only the Ethernet interfaces that belong to the same card are only
interfaces belong to the same bond interface.
Any change of bonding configuration (i.e, bonding mode is not 1, 2 or 4,
or both Ethernet interfaces that belong to the same card are not the
only interfaces belong to the bond interface), RoCE bond becomes
inactive.

When RoCE bond is active, only one IB device is been registered(uses
pf0 to register), and both port0 and port1 need to transmit traffics
for this IB device, for load-balance or active-backup.

Main changes in this patch contain:
1. alloc a `crdma_bond` for each card, it services all `nfp_roce` on
one card, it create when first `nfp_roce` created and destroy when last
`nfp_roce` destroyed;
2. monitor the bond interface's netdev events,  record and check whether
RoCE bond becomes active or not;
3. update RoCE bond status to firmware.

Signed-off-by: Li Xu <li.xu@corigine.com>
Signed-off-by: Xiaofeng Tang <xiaofeng.tang@corigine.com>
Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>