nfp: add support for cpp cdev to use area cache
In our currently cpp implement, there are more and more reserved bars
in device probe stage. There are some cpp area caches which will be
used to keep the cpp area and bar when the cpp access is finished for
the following access to use. It is effective in most of the cases,
but in our current implement, the cpp access via cdev will not use
area cache which will lead the cpp access via cdev will not get the
bars hold by area cache.

When some user commands depend on some specified bars and the bars
are reserved or hold by area cache, the command will hang until the
require bar is released or replaced in cache.

So in this implement, we add support for cpp cdev to use area cache,
then the user command will not hang if the required bar is hold in
area cache.

Signed-off-by: Baowen Zheng <baowen.zheng@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>