nfp: flower: reject to offload flow with empty action
Our current flower implement can offload a flow with empty action and
the packets match this flow will be dropped.

Actually in this case, packets should fallback to host to allow host
process these packets, this is the behavior in software datapath. So we
decide to reject to offload flow with empty action since anyway packets
will fallback to host.

Signed-off-by: Baowen Zheng <baowen.zheng@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>