workflows: sync-local-kernels: force use of azcopy vs local mount
When this workflow is called by the "sync-local-buildinfo" workflow,
that workflow will have just uploaded new files to Azure, these will not
yet have synced to the local mount, and so the "sync-local-kernels"
workflow will fail.

This change will force the "azcopy_download" step of the workflow to
use azcopy instead of copying files from the local mount.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>