workflows: build-local: omit jobs when OS not in data
Add a step to the prepare job of the build-local.yaml workflow to not
add certain operating systems to the build matrix if there is no
buildinfo.json file for that operating system in
.github/data/kernel-devel-c4c/.

It is assumed that buildinfo files for a distro will only be added to
these directories when there are appropriate compat patches already
present.

This will ensure that we do not attempt to build on distributions when
there is no compat code for the driver version being built.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>