workflows: build: limit download ranges for Ubuntu 20.40 and 22.04
There are > 100 kernel build-versions available for the kernels on
which these two OS's were released (v5.4 and v5.15 respectively). It is
unnecessary to test on all of them, so limit the range to only the
newer versions. Also remove some of the older hwe kernel versions for
Ubuntu 20.04.

- Ubuntu 20.04, remove 5.8, 5.11 and 5.13 from the list
- Ubuntu 20.04, limit to build versions > 150
- Ubuntu 22.04, limit to build versions > 100

Also add the 'hwe' flag to the job name, so it is easier to identify
the runs.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>