nfp: fix build error of unexpected BPF reg types
Some C4C kernels like openEuler and derived Anolis modified the
definition of BPF reg types in its 5.10.x version somehow, which never
happened in upstream kernel, and this causes compiling error.

The original change fixes the compiling error but breaks the
functionality. Now return unsupported if the definitions are not as
expected.

Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>