nfp: crdma: expand sgid table size to 16
Previous size is 4, too small.

Currently in bond active-backup mode, if the active link switches,
the gid table will be updated and reissued. If a GID has been in
use during the link switch, then the entry corresponding to the GID
is equivalent to being occupied. The gid table update needs to use
another unoccupied entry in the table to reissued GIDs.

Therefore, if two IPv4 addresses are currently configured, the gid
table already has 3 GIDs then. If the GIDs corresponding to the two
IPv4 addresses are being used, then when the above situation occurs,
the GIDs will be reissued, and one would lost due to lack of space.

Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Li Xu <li.xu@corigine.com>
Signed-off-by: Xiaofeng Tang <xiaofeng.tang@corigine.com>
Signed-off-by: Hui Zhang <hui.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>