workflows: update azcopy provisioning to use static link
Update instances across all workflows in the repo which download and
install the azcopy binary to make use of the static link to release
v10.24.0, as apposed to the dynamic link.

The dynamic link redirects to v10.25.0 currently and this version
introduces a bug which causes our usage of the "azcopy list" command to
return an empty string.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>