workflows: build: update CentOS7 repository mirror
CentOS7 reached end-of-life, and the original mirror have been
discontinued. It still lives on at 'vault.centos.org', so use sed to
update the repo URL's to point to this instead.

The download URL for CentOS 7.9 now also moved to 'vault.centos.org',
so add this minor modification as well.

Signed-off-by: Louis Peens <louis.peens@corigine.com>