tools: add script to check for if crdma should be compiled
In order for the dkms packaging to figure out if 'crdma.ko' is expected
to be compiled some helpers are needed. This script duplicates all the
checks for BUILD_CRDMA from the Kbuild file. It is unfortunate that
these checks have to be duplicated, but this seems to be the best
solution at this point in time.

This script can be run standalone, with input arguments being the
kernel version, and the path to the kernel headers, e.g

./check_crdma_sup.sh 5.14.0-427.22.1.el9_4.x86_64 \
    /lib/modules/5.14.0-427.22.1.el9_4.x86_64/build

This will output either '0', or 'y', depending on whether or not CRDMA
can be compiled on the system - at least according to kernel
requirements.

The script is mainly designed to work with the dkms tool, which exposes
the $kernelver and $kernel_source_dir variables. This script does very
little input validation, as it assumes that dkms is already providing
valid inputs. Hooking this into the dkms packaging path will follow
soon.

If the script is executed with the CRDMA_DEBUG variable set it will
print the derived variables for debugging.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Fei Qin <fei.qin@corigine.com>