packaging: dkms: use check_crdma_sup.sh script to enable crdma build
Use the newly added check_crdma_sup.sh script to determine if the dkms
packaging should package the 'crdma.ko' module. This calls the script
from the source tree of the specific dkms module, giving it the
$kernelver and $kernel_source_dir arguments, which are provided by the
dkms subsystem.

If this returns "y" then the module is added, otherwise just the
'nfp.ko' module will be compiled.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Fei Qin <fei.qin@corigine.com>