workflows: add a short term workaround for Node16 EOL
Github actions recently switched over to only running on Node20. This
is a problem for CentOS 7, as the github provided Node20 does not work
in CentOS 7. There is a short term workaround, by adding
    ACTIONS_ALLOW_USE_UNSECURE_NODE_VERSION: true
to the job's environment variable will allow it to still use Node16 as
long as it is on the system.

Apply this while a proper fix is being investigated.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>