nfp: don't update coalesce params with adaptive coalesce enabled
Currently, the coalesce parameters can be changed with adaptive
coalesce enabled. And the new configured parameters can take effect to
influence the configuration set by dim algorithm, which looks not quite
reasonable.

This patch avoids the coalesce params to be configured to firmware when
adaptive coalesce is enabled. But the new configurations still can be
stored and take effect when the adaptive coalesce is turned off.

Signed-off-by: Fei Qin <fei.qin@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>