nfp: crdma: Prevent doorbell lossing
When the driver knocks the doorbell at a burst frequency, the hardware
queue may be full and drop the doorbell due to untimely acquisition or
limited acquisition capabilities. So we use 2 ME and 8 context for each
ME to fetch the doorbells.
To solve the sequencing issue, for each type of doorbell, including
SQ/RQ/CQ/EQ, we always ensure that the latest request is processed
while older requests are ignored. Therefore, we introduce a serial
number to handle the comparison of sequence after data reversing.

Signed-off-by: Xinxia Qu <xinxia.qu@corigine.com>
Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Signed-off-by: Xiaofeng Tang <xiaofeng.tang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>