nfp: adjust DIM profile for latency testing
When DIM is enabled, for latency testing such as qperf/netperf/ping,
the running interrupt coalesce parameters rx-usecs and rx-frames
will be switched between {0,1} and {4,32} which is defined by
DIM profile. However, according to the test result, it is obvious
that the latency of packets is lower when packets are not coalesced.
That means the latency is better with usecs = 0 or frames = 1.

In order to improve latency in such situation, it is better to merge
the entry 0 and 1 in the DIM profile. So that for latency testing,
such as qperf/netperf/ping, the packets are generally not coalesced.

Before this patch:
qperf: 17-21us
netperf: 34-40us
ping: 40-50us

After this patch:
qperf: 16-18us
netperf: 33-35us
ping: 30-40us

Signed-off-by: Rick Zhong <zhaoyong.zhong@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>