nfp: sgw: add vnic init and clean
Expand the existing flower vnic functions to work correctly when
the app is sgw. Since flower and sgw vnic setup is very similar
it is better to avoid duplication by only taking a different path
when needed. Add a 'nfp_app_is_sgw' helper function, and use this
in the few places where the behaviour is different between sgw
and flower.

Signed-off-by: xuehao yin <xuehao.yin@corigine.com>
Signed-off-by: Kang Pan <kang.pan@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>