nfp: sgw: add configurable queue number config
For sgw the representor ports can have a configurable number of rx
and tx rings. This is different to flower which just multiplexes
everything over the PF vnic queues. This patch starts introducing
this feature by adding the relevant fields to various structures.

There are various different NIC variants available in use for sgw,
each with different port numbers and queue configurations.
This patch adds a static configuration mapping for each of these
card variants, based on the values read from the 'partno'
field of nfp-hwinfo.

Signed-off-by: xuehao yin <xuehao.yin@corigine.com>
Signed-off-by: Kang Pan <kang.pan@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>