tools: nfp_troubleshoot_gather: ensure script cleans up
Previously the script could fail if something went wrong (such as a BSP
tool failure or unexpected command output), resulting in the
intermediate log directory not being tarred or deleted.

This patch will not resolve any underlying issues, but instead aims to
guarantee that the tar file is always created even if the information
gathered is incomplete.

Signed-off-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>