tools: troubleshoot: always log stderr to script_cmds.log
There are cases where commands are failing during the gathering
process, but the reason is not clear. When fail=True is passed to
scmd() this is not a problem, since both STDERR and STDOUT will be
captured. However - when fail=False this will not happen, and from the
'scripts_cmds.log' file it is clear in those cases that the return code
was non-zero, but the reason is not clear.

Get around this problem by always logging STDERR, regardless of the
'fail' flag. Since STDERR is usually quite short adding it to the log
is not expected to expand the logfile size too much.

Capturing STDOUT is avoided for now, since this will lead to a lot of
duplicated information (in the script log and the specific tools log).
If the debug information is still not good enough after this further
updates can be made to also capture STDOUT in specific cases, or in a
specific debug mode.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>