nfp: support the lldp dcbx packet parse
This patch adds support for reading dcbx config through configuration
messages received in LLDP packets. It adds a new 'NFP_NET_META_LLDP'
metadata type, which is used as an indicator that the packet should
be parsed as LLDP.

A new 'lldp_packet_parse' application callback is added to allow
parsing of these LLDP packets, and one implementation is added in the
'nic' app. This implements code to parse all the different DCBx configs
from the TLV fields found in the LLDP packet. This patch only parses
the LLDP packets and allows dumping the information - it does not
configure DCB on the card yet.

Lastly a new 'enable-rx-lldp' flag is added to the ethtool
'--show/set-priv-flags' to enable/disable the feature.

Signed-off-by: Weiwen Hu <weiwen.hu@corigine.com>
Reviewed-by: Bin Chen <bin.chen@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>