nfp: support the lldp rx configuration to save into hw-info
This patch fixes three issues. The first one is to solidify the lldp-rx
configuration to avoid loss during restart. The second one is to set
the lldp tx switch name to be more user-friendly. The 'disable-fw-lldp'
flag is being removed in favour of the more descriptive
'enable-tx-lldp' flag, to match the recently added 'enable-rx-lldp'
flag.

Signed-off-by: Weiwen Hu <weiwen.hu@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Yong Xu <yong.xu@corigine.com>
Reviewed-by: Bin Chen <bin.chen@corigine.com>