tools: corigine_pciids: Temporarily workaround to fix lspci breakage
The optimised path for when entries are already present breaks lspci.
If the vendor is already present but some of the device entries are not
then they will be added at the end instead of at their appropriate
location in the file:
	C ff  Unassigned class

	# Next line added by agilio-nfp-driver package
		4000  Network Flow Processor 4000
	# Next line added by agilio-nfp-driver package
		6003  Network Flow Processor 6003 Virtual Function

Fix the behaviour at the cost of performance by forcing the
optimisation flag to false for now.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>