tools: corigine_pciids.sh: fix script
Previously, the script would incorrectly apply update lines to the end
of the target file if more than one of the update lines were already
present in the target file AND these lines were adjacent.

This happened because `grep` would report a line is present in the
file when that line had already been consumed by a previous read. This
means that the subsequent `sed` read would never encounter that line and
stop, so it just read to the end of the file. A fool-proof solution
would be to keep track of which line number has last been read and only
grep from that line onwards. But doing so would significantly slow down
the script when the optimisation cannot be applied.

Instead, use the fact that both the update contents (from the
here-string) and the target file will always contain entries in the
correct order. So it is only necessary to ensure that the ID is later
than the last read line. (Because if the match is earlier in the file,
it means the original contents was invalid; and if the match is later
in the file, the `sed` optimisation will work as intended.)

Fixes: 22f96bfa3573 ("tools: add a script to update pci ids")
Signed-off-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>