tools: corigine_pciids.sh: try to prevent permission issues
Create the temporary files in `/var/run` instead of `/tmp` due to write
permissions of certain users.

Also add a timestamp to temporary files to avoid permission errors when
two different users try to run the script.

Also, do not overwrite the backup file if one already exists. This
avoids losing the backup if `apply` is run multiple times.

Signed-off-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>